Extract automation v2 models into crate
