+++ crates/tandem-automation/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "tandem-automation"
+version = "0.6.1"
+description = "Automation model and execution primitives for Tandem"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/frumu-ai/tandem"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tandem-enterprise-contract = { path = "../tandem-enterprise-contract", version = "0.6.1" }
+tandem-orchestrator = { path = "../tandem-orchestrator", version = "0.6.1" }
+tandem-plan-compiler = { path = "../tandem-plan-compiler", version = "0.6.1" }
+tandem-types = { path = "../tandem-types", version = "0.6.1" }
+tandem-workflows = { path = "../tandem-workflows", version = "0.6.1" }
