+++ crates/tandem-automation/src/routine.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case", tag = "type")]
+pub enum RoutineMisfirePolicy {
+    Skip,
+    RunOnce,
+    CatchUp { max_runs: u32 },
+}
