+++ crates/tandem-server/Cargo.toml
@@ -51,6 +51,7 @@ tandem-core = { path = "../tandem-core", version = "0.6.1" }
 tandem-providers = { path = "../tandem-providers", version = "0.6.1" }
 tandem-runtime = { path = "../tandem-runtime", version = "0.6.1" }
 tandem-tools = { path = "../tandem-tools", version = "0.6.1" }
+tandem-automation = { path = "../tandem-automation", version = "0.6.1" }
 tandem-skills = { path = "../tandem-skills", version = "0.6.1" }
 tandem-memory = { path = "../tandem-memory", version = "0.6.1" }
 tandem-orchestrator = { path = "../tandem-orchestrator", version = "0.6.1" }
